+++ services/server/test/unit/verificationWorker.spec.ts
@@ -9,7 +9,7 @@ import {
 } from "../../src/server/services/workers/verificationWorker";
 import Sinon from "sinon";
 import Piscina from "piscina";
-import { sourcifyChainsMap } from "../../src/sourcify-chains";
+import { LOCAL_CHAINS } from "../../src/sourcify-chains";
 import {
   SolidityCompilation,
   type SourcifyChainInstance,
@@ -35,10 +35,10 @@ describe("verificationWorker", function () {
   const chainFixture = new LocalChainFixture();
   const piscinaSandbox = Sinon.createSandbox();
 
-  before(() => {
-    const sourcifyChainInstanceMap = Object.entries(sourcifyChainsMap).reduce(
-      (acc, [chainId, chain]) => {
-        acc[chainId] = chain.getSourcifyChainObj();
+  before(async () => {
+    const sourcifyChainInstanceMap = LOCAL_CHAINS.reduce(
+      (acc, chain) => {
+        acc[chain.chainId.toString()] = chain.getSourcifyChainObj();
         return acc;
       },
       {} as Record<string, SourcifyChainInstance>,