+++ services/4byte/CHANGELOG.md
@@ -1,4 +1,6 @@
+## sourcify-4byte@1.1.12 - 2026-05-21
 
+- Update dependencies
 
 ## sourcify-4byte@1.1.11 - 2026-04-14