+++ services/monitor/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## sourcify-monitor@1.5.6 - 2026-05-21
+
+- fix(server): throw on chains with no usable RPCs, add DRPC_API_KEY
+  #2782
+- Use @ethereum-sourcify/test-helpers
+- Update dependencies
+
 ## sourcify-monitor@1.5.5 - 2026-05-06
 
 - update dependencies